fix: static build
